Rename files and remove unnecessary comments
