+++ src/services/example-service/client.ts
@@ -5,30 +5,20 @@ import {
   Response,
   ResponsePromise,
   SuccessResponse
-} from 'src/services/example-service/service-api-responses';
+} from 'src/services/example-service/types';
 
-/**
- * Number of attempts before request fails
- */
 const MAX_RETRY_ATTEMPTS = 3;
 
-/**
- * Axios client for talking with a service
- */
 export const axiosClient = axios.create({
   baseURL: process.env.SERVICE_URL,
   responseType: 'json',
   validateStatus: () => true
 })
 
-/**
- * Wrapper function to make a request to the service
- */
 export const makeRequest = async<T>(
   config: AxiosRequestConfig,
   attempts = 1
 ): ResponsePromise<T> => {
-  // Merge the Authorization header with any other potential headers
   config.headers = {
     ...(config.headers || {}),
     ...{
@@ -38,24 +28,13 @@ export const makeRequest = async<T>(
 
   const axiosResponse = await axiosClient.request<T>(config);
 
-  // Retry the request until we've reached the maximum attempts
-  if (
-    !responseIsSuccessful(axiosResponse) &&
-    attempts < MAX_RETRY_ATTEMPTS
-  ) {
+  if (!responseIsSuccessful(axiosResponse) && attempts < MAX_RETRY_ATTEMPTS) {
     return makeRequest<T>(config, attempts + 1);
   }
 
   return buildApiResponse(axiosResponse, config);
 }
 
-/**
- * Construct an api response based off of a response returned
- * from Axios
- *
- * @param response
- * @param initialConfig
- */
 const buildApiResponse = <T>(
   response: AxiosResponse<T>,
   initialConfig: AxiosRequestConfig
@@ -89,12 +68,6 @@ const buildApiResponse = <T>(
   return successResponse;
 }
 
-/**
- * Utility class to determine whether or not a response was
- * successful
- *
- * @param response
- */
 const responseIsSuccessful = (response: AxiosResponse): boolean => {
   if (response.status < 200 || response.status >= 300) {
     return false;
