+++ src/services/example-service/index.ts
@@ -1,0 +1,3 @@
+export * from './client';
+export * from './getExample';
+export * from './types';
