--- src/services/example-service/service-types.d.ts
@@ -1,3 +1,0 @@
-export type Example = {
-  name: string
-}
