+++ src/services/todo/api.ts
@@ -1,5 +1,5 @@
 import { nanoid } from '@reduxjs/toolkit';
-import { Todo } from './todo-api-types';
+import { Todo } from './types';
 
 const todos: Todo[] = [
   {
@@ -9,16 +9,10 @@ const todos: Todo[] = [
   },
 ];
 
-/**
- * Gets the list of todos
- */
 export async function listTodos(): Promise<Todo[]> {
   return todos;
 }
 
-/**
- * Creates a new todo item
- */
 export async function createTodo(content: string): Promise<Todo> {
   if (content === 'Delay') {
     await new Promise(r => setTimeout(r, 2000));
